# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: UltraV API
  version: 1.0.0
  summary: API-first KVM/QEMU virtual machine management
  description: |
    UltraV is a modern, simple, API-first management layer for KVM/QEMU
    virtual machines on single Linux hosts.

    **Everything is an API.** Every UI capability is available here first.
    This contract is the source of truth for the Web UI, CLI, scripts,
    automation, and future AI agents (MCP adapters).

    ## Conventions
    - All requests/responses are JSON.
    - Every response includes an `X-Request-Id` header; errors also include `requestId` in the body.
    - Errors use a stable envelope: `{"error": {"code", "message", "requestId"}}`.
    - Destructive or long-running operations will use specific, semantically
      named operations (never generic `execute`/`command` endpoints).
  license:
    name: TBD
servers:
  - url: /api/v1
    description: Local host (single-node)
...
paths:
...
  /vms/{id}:
    get:
      operationId: getVirtualMachine
      tags: [Virtual Machines]
      summary: Get a virtual machine
      parameters: [ { $ref: '#/components/parameters/VMId' } ]
      responses:
        '200':
          description: Virtual machine details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VirtualMachine' }
        '404': { $ref: '#/components/responses/VMNotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    patch:
      operationId: updateVirtualMachine
      tags: [Virtual Machines]
      summary: Update virtual machine settings
      description: >
        Changes the VM's hardware configuration: vCPU count, memory size and
        the attached ISO (install media). Only the fields present in the body
        are changed (partial update). vCPU and memory changes require the VM
        to be stopped; the ISO can be swapped in any state and takes effect on
        the next boot.
      parameters: [ { $ref: '#/components/parameters/VMId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VirtualMachineUpdate' }
            example: { vcpus: 4, memoryBytes: 8589934592, isoId: '' }
      responses:
        '200':
          description: Updated virtual machine
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VirtualMachine' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/VMNotFound' }
        '409': { $ref: '#/components/responses/VMInvalidState' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      operationId: deleteVirtualMachine
      tags: [Virtual Machines]
      summary: Delete a virtual machine
      description: |
        Removes a virtual machine. Two levels of deletion are supported via
        the deleteDisks query parameter:

        * `deleteDisks=false` (default): undefines only the domain
          configuration. The disk volumes remain in the storage pool and can
          be reclaimed or reused manually.
        * `deleteDisks=true`: undefines the domain AND deletes its disk
          volumes from the storage pool, freeing the space.

        The VM must be stopped (409 otherwise).
      parameters:
        - { $ref: '#/components/parameters/VMId' }
        - name: deleteDisks
          in: query
          required: false
          schema: { type: boolean, default: false }
          description: Also delete the VM's disk volumes from the storage pool.
      responses:
        '204': { description: Virtual machine deleted }
        '404': { $ref: '#/components/responses/VMNotFound' }
        '409': { $ref: '#/components/responses/VMInvalidState' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /vms/{id}/backups:
    get:
      operationId: listVMBackups
      tags: [Backup]
      summary: List backups of a virtual machine
      parameters: [ { $ref: '#/components/parameters/VMId' } ]
      responses:
        '200':
          description: Backup points of this VM (newest first)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BackupList' }
        '404': { $ref: '#/components/responses/VMNotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: createVMBackup
      tags: [Backup]
      summary: Create a backup of a virtual machine (full or incremental)
      description: |
        Backs up every disk of the VM plus its domain configuration.

        By default the engine picks the point type automatically:
        **incremental** when a valid backup chain exists (previous point's
        checkpoint present in the domain — only dirty blocks since the last
        backup are copied, via libvirt checkpoints / QEMU dirty bitmaps),
        otherwise **full**. Pass `{"type": "full"}` to force a full backup
        and reset the chain.

        Works with the VM running or stopped. When running, the backup is
        crash-consistent (block-level snapshot taken by the hypervisor via
        `virDomainBackupBegin`); for application-consistent backups of
        databases, shut the VM down or quiesce it first (guest-agent fsfreeze
        integration arrives in a later phase). An offline (stopped) backup is
        always full.

        The call is synchronous: it returns when the copy completes. Disk
        images are stored as qcow2 in the server's backup directory
        (`ULTRAV_BACKUP_DIR`, default `/var/lib/ultrav/backups`), one folder
        per backup point, alongside a `metadata.json` and the saved
        `domain.xml`.
      parameters: [ { $ref: '#/components/parameters/VMId' } ]
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BackupCreate' }
            example: { type: incremental }
      responses:
        '202':
          description: |
            Backup accepted and running in the background as a task. Poll
            `GET /tasks/{id}` until the status is `succeeded` or `failed`;
            the finished task's `resourceId` is the backup point id.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '404': { $ref: '#/components/responses/VMNotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /backups/{id}:
    delete:
      operationId: deleteBackup
      tags: [Backup]
      summary: Delete a backup point
      description: Permanently removes the backup point's directory (disk images, metadata and saved domain XML). Does not touch the VM.
      parameters:
        - name: id
          in: path
          required: true
          description: Backup identifier (`bk-...`).
          schema: { type: string, pattern: '^bk-[a-zA-Z0-9._-]{1,80}$', example: bk-20250901-120000-a1b2 }
      responses:
        '204': { description: Backup deleted }
        '404': { $ref: '#/components/responses/BackupNotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /backups/{id}/restore:
    post:
      operationId: restoreBackup
      tags: [Backup]
      summary: Restore a backup into its virtual machine
      description: |
        Overwrites the disks of the VM the backup was taken from with the
        backup's disk images, and restores the saved domain configuration.
        Incremental points are restored by replaying the whole chain
        (full → incrementals, oldest first). The VM must exist and be stopped
        (409 otherwise) — restore is destructive: the current disk contents
        are replaced.

        After a restore the backup chain is reset: the next backup will be a
        new full (the domain's dirty-bitmap checkpoints are deleted, since
        the disks now reflect an older point in time).

        The restored VM reflects the state captured by the backup (hardware
        settings come from the saved `domain.xml`). Backups of VMs that no
        longer exist cannot be restored in this phase (delete the old VM
        metadata manually or wait for restore-as-new-VM).
      parameters:
        - name: id
          in: path
          required: true
          description: Backup identifier (`bk-...`).
          schema: { type: string, pattern: '^bk-[a-zA-Z0-9._-]{1,80}$', example: bk-20250901-120000-a1b2 }
      responses:
        '202':
          description: |
            Restore accepted and running in the background as a task. Poll
            `GET /tasks/{id}` until the status is `succeeded` or `failed`.
            Failure reasons (VM running, backup's VM no longer exists) are
            reported in the task's `error`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '404': { $ref: '#/components/responses/BackupNotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /vms/{id}/config-export:
    get:
      operationId: exportVMConfig
      tags: [Backup]
      summary: Export the VM configuration (domain XML)
      description: Returns the libvirt domain XML of the VM, for configuration backup / disaster recovery. Cheap to call; pair it with backups for full-VM restore on a blank host.
      parameters: [ { $ref: '#/components/parameters/VMId' } ]
      responses:
        '200':
          description: Domain XML (content type application/xml)
          content:
            application/xml:
              schema: { type: string }
        '404': { $ref: '#/components/responses/VMNotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
security:
  - bearerAuth: []
components:
  parameters:
...
    VMId:
      name: id
      in: path
      required: true
      description: Virtual machine identifier (name).
      schema: { type: string, pattern: '^[a-zA-Z0-9][a-zA-Z0-9._-]{0,63}$', example: erp01 }
...
  responses:
    VMNotFound:
      description: Virtual machine not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: VM_NOT_FOUND, message: Virtual machine was not found, requestId: req_01HV3M9Z2K }
    VMInvalidState:
      description: Operation not valid for the current VM state
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: VM_INVALID_STATE, message: Virtual machine is not running, requestId: req_01HV3M9Z2L }
    BackupNotFound:
      description: Backup point not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: BACKUP_NOT_FOUND, message: Backup was not found, requestId: req_x } }
...
    InternalError:
      description: Internal server error (no internal details are exposed)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: INTERNAL_ERROR, message: An internal error occurred, requestId: req_01HV3M9Z2M }
...
    ValidationError:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: VALIDATION_ERROR, message: username and password are required, requestId: req_x } }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
              example: VM_NOT_FOUND
              description: >
                Stable error identifier. Known values: VM_NOT_FOUND, VM_INVALID_STATE,
                VM_ALREADY_EXISTS, VALIDATION_ERROR, NOT_FOUND, METHOD_NOT_ALLOWED,
                STORAGE_POOL_NOT_FOUND, STORAGE_POOL_ALREADY_EXISTS, NETWORK_NOT_FOUND,
                NETWORK_INVALID_STATE, NETWORK_INACTIVE (a virtual network a VM references
                is not active — start it before powering on the VM), NETWORK_ALREADY_EXISTS,
                ISO_NOT_FOUND, ISO_ALREADY_EXISTS, UNAUTHORIZED, INVALID_CREDENTIALS,
                FORBIDDEN, USER_NOT_FOUND, USER_ALREADY_EXISTS, LAST_ADMIN,
                CONSOLE_UNAVAILABLE, STORAGE_UNAVAILABLE (the hypervisor cannot open a file
                the domain references), KVM_UNAVAILABLE (hardware virtualization missing or
                inaccessible on the host), BACKUP_NOT_FOUND, BACKUP_INVALID_STATE (the backup's
                VM is running or no longer exists — stop the VM first), SCHEDULE_NOT_FOUND,
                TASK_NOT_FOUND, TASK_NOT_CANCELLABLE (the task already reached a
                terminal state), INTERNAL_ERROR.
            message: { type: string, example: Virtual machine was not found }
            requestId: { type: string, example: req_01HV3M9Z2K }
...
    VMState:
      type: string
      enum: [running, stopped, paused, shutting-down, starting, error]
      description: |
        - running: powered on and operational
        - stopped: powered off
        - shutting-down: graceful shutdown in progress
        - starting: power-on in progress
...
    VirtualMachineUpdate:
      type: object
      description: >
        Partial update of a VM's settings. Only provided fields are changed.
        vCPU and memory require the VM to be stopped. For isoId, a filename
        attaches/swaps the install media and an empty string detaches it.
      properties:
        vcpus:
          type: integer
          minimum: 1
          maximum: 64
          example: 4
        memoryBytes:
          type: integer
          format: int64
          minimum: 16777216
          example: 8589934592
        isoId:
          type: string
          description: ISO filename from the library to attach/swap; an empty string detaches the current ISO.
          example: ubuntu-24.04-live-server.iso
        networkId:
          type: string
          description: |
            Moves the VM's network interface to another virtual network.
            Requires the VM to be stopped (applied to the persistent config).
          example: br0
...
    Backup:
      type: object
      required: [id, vmId, vmName, type, createdAt, sizeBytes, disks]
      properties:
        id: { type: string, example: bk-20250901-120000-a1b2 }
        vmId: { type: string, example: erp01 }
        vmName: { type: string, example: erp01 }
        type:
          type: string
          enum: [full, incremental]
          example: incremental
          description: |
            `full` copies every disk in full and resets the chain;
            `incremental` copies only the blocks modified since the previous
            point (requires a valid chain, i.e. the parent's checkpoint still
            exists in the domain).
        parentId:
          type: string
          nullable: true
          description: For incremental points, the id of the backup this one continues from.
          example: bk-20250831-120000-c3d4
        state: { type: string, enum: [complete], example: complete }
        createdAt:
          type: string
          format: date-time
          example: '2025-09-01T12:00:00Z'
        sizeBytes:
          type: integer
          format: int64
          description: Total size of the backup's disk images on the backup storage.
          example: 34359738368
        disks:
          type: array
          items: { $ref: '#/components/schemas/BackupDisk' }
        hasDomainXml: { type: boolean, example: true }
    BackupCreate:
      type: object
      properties:
        type:
          type: string
          enum: [full, incremental]
          description: |
            Force the point type. `incremental` fails with 409
            `BACKUP_INVALID_STATE` when no valid chain exists (e.g. after a
            restore or if the parent checkpoint was deleted). Absent = auto
            (incremental when a chain is valid, full otherwise).
          example: incremental
...
    BackupDisk:
      type: object
      required: [name, file, sizeBytes]
      properties:
        name: { type: string, example: vda }
        file: { type: string, example: vda.qcow2 }
        format: { type: string, enum: [qcow2, raw], example: qcow2 }
        sizeBytes: { type: integer, format: int64, example: 34359738368 }
    BackupList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Backup' }
        total: { type: integer, example: 3 }
    Disk:
      type: object
      required: [name, format, sizeBytes]
      properties:
        name: { type: string, example: vda }
        format: { type: string, enum: [qcow2, raw], example: qcow2 }
        sizeBytes: { type: integer, format: int64, example: 128849018880 }
        usedBytes:
          type: integer
          format: int64
          nullable: true
          description: >
            Space currently allocated for this disk on the host (the backing
            volume's allocation, not the guest filesystem usage). Null when the
            hypervisor cannot report it.
          example: 34359738368
        bus: { type: string, enum: [virtio, sata, scsi], example: virtio }
    VmMetrics:
      type: object
      description: >
        Point-in-time resource utilization of a running virtual machine. The
        whole object is absent while the VM is not running. Individual fields
        are null when the host cannot report them (e.g. no balloon driver /
        guest agent, or a first sample with no previous sample to diff
        against). Values are computed by the hypervisor provider, which is
        allowed to keep a short-lived sampling cache per VM.
      required:
        [cpuPercent, memoryUsedBytes, networkRxBytesPerSecond, networkTxBytesPerSecond, sampledAt]
      properties:
        cpuPercent:
          type: number
          format: float
          nullable: true
          minimum: 0
          maximum: 100
          description: CPU utilization relative to the VM's allocated vCPUs.
          example: 12.5
        memoryUsedBytes:
          type: integer
          format: int64
          nullable: true
          description: >
            Guest memory currently in use. Null unless the guest balloon
            driver or agent reports it.
          example: 3221225472
        networkRxBytesPerSecond:
          type: integer
          format: int64
          nullable: true
          description: >
            Inbound network throughput across all interfaces. Null until a
            second sample is available to compute a rate.
          example: 40960
        networkTxBytesPerSecond:
          type: integer
          format: int64
          nullable: true
          description: >
            Outbound network throughput across all interfaces. Null until a
            second sample is available to compute a rate.
          example: 20480
        sampledAt:
          type: string
          format: date-time
          description: When the underlying sample was taken.
          example: '2025-09-01T12:00:00Z'
    NetworkInterface:
      type: object
      required: [name, model]
      properties:
        name: { type: string, example: ens3 }
        model: { type: string, enum: [virtio, e1000, rtl8139], example: virtio }
        macAddress: { type: string, example: '52:54:00:a1:b2:c3' }
        ipAddress: { type: string, example: 10.0.0.11 }
        network: { type: string, example: default }
    VirtualMachine:
      type: object
      required: [id, name, state, vcpus, memoryBytes, uptimeSeconds, disks, networkInterfaces]
      properties:
        id: { type: string, example: erp01 }
        name: { type: string, example: erp01 }
        state: { $ref: '#/components/schemas/VMState' }
        vcpus: { type: integer, example: 4 }
        memoryBytes: { type: integer, format: int64, example: 8589934592 }
        uptimeSeconds: { type: integer, nullable: true, example: 345600 }
        os: { type: string, example: Ubuntu 24.04 LTS }
        ipAddress: { type: string, nullable: true, example: 10.0.0.11 }
        isoId:
          type: string
          nullable: true
          description: Filename of the ISO currently attached as install media (CD-ROM), if any.
          example: ubuntu-24.04-live-server.iso
        disks:
          type: array
          items: { $ref: '#/components/schemas/Disk' }
        networkInterfaces:
          type: array
          items: { $ref: '#/components/schemas/NetworkInterface' }
        metrics: { $ref: '#/components/schemas/VmMetrics' }
        bootTime:
          type: string
          format: date-time
          nullable: true
          example: '2025-09-01T12:00:00Z'
        osType:
          type: string
          enum: [linux, windows, other]
          nullable: true
          description: |
            Guest OS family detected from the domain XML (presence of Hyper-V
            enlightenments implies `windows`; VMs created before profiles
            existed may report `null`).
          example: windows
        performanceProfile:
          $ref: '#/components/schemas/PerformanceProfile'
        warnings:
          type: array
          items: { type: string }
          nullable: true
          description: |
            Non-fatal notices, populated only in the create response (e.g. a
            Windows VM created without a VirtIO drivers ISO). Absent/null on
            reads.
          example:
            - 'Windows without a VirtIO drivers ISO: the installer will not see the virtio-scsi disk until vioscsi is loaded from another source.'
    PerformanceProfile:
      type: object
      description: |
        Performance-oriented settings applied to the domain XML at creation,
        as detected from the domain (null for VMs created before profiles
        existed).
      properties:
        cpuMode:
          type: string
          enum: [host-passthrough, host-model, custom]
          description: '`<cpu mode>` applied. host-passthrough gives best performance but limits live migration between hosts with different CPUs (multi-host phase concern).'
          example: host-passthrough
        diskBus:
          type: string
          enum: [scsi, sata]
          description: Disk bus applied (virtio-scsi for linux/windows profiles, SATA for other).
          example: scsi
        cache:
          type: string
          enum: [none]
          description: Disk cache mode applied.
          example: none
        ioThreads:
          type: integer
          nullable: true
          description: Number of iothreads dedicated to the SCSI controller (0/absent when not applied, e.g. the `other` profile).
          example: 1
        nicModel:
          type: string
          enum: [virtio, e1000e, rtl8139]
          nullable: true
          description: Network adapter model applied (null for VMs created before this was recorded).
          example: virtio
        hypervEnlightenments:
          type: array
          items: { type: string }
          description: |
            Hyper-V enlightenments actually present in the domain XML (only
            for `windows`; only the ones supported by the host's
            QEMU/libvirt, as detected from domain capabilities).
          example: [relaxed, vapic, spinlocks, vpindex, synic, stimer]
    TaskStatus:
      type: string
      enum: [queued, running, succeeded, failed, cancelling, cancelled]
      description: >
        queued: accepted, not started yet. running: worker executing.
        cancelling: cancel requested, waiting for the worker to return.
        succeeded/failed/cancelled: terminal states.
    Task:
      type: object
      required: [id, type, status, progress, cancellable, createdAt]
      properties:
        id: { type: string, example: task-000001 }
        type:
          type: string
          enum: [vm-create, backup-create, backup-restore]
          example: vm-create
        status: { $ref: '#/components/schemas/TaskStatus' }
        resourceId:
          type: string
          description: Identifier of the produced resource (VM name, backup point id). Empty while unknown.
          example: app01
        resourceName:
          type: string
          description: Human-readable target of the task (e.g. the VM name).
          example: app01
        message:
          type: string
          description: Current step description, when available.
          example: allocating disk volume
        error:
          type: string
          description: Failure reason when status is failed.
        progress:
          type: integer
          minimum: 0
          maximum: 100
          example: 40
        warnings:
          type: array
          description: Non-fatal notes collected during execution (e.g. profile warnings).
          items: { type: string }
        cancellable:
          type: boolean
          description: Whether cancelTask can still be called.
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
...